+++ app/src/main/java/com/example/b07group7project/nav/ShopperNavigationActivity.java
@@ -6,6 +6,7 @@ import android.os.Bundle;
 
 import com.example.b07group7project.ExampleFragment;
 import com.example.b07group7project.R;
+import com.example.b07group7project.create_product.CreateProductFragment;
 import com.example.b07group7project.databinding.ShopperNavigationActivityBinding;
 import com.example.b07group7project.itempreview.ItemPreviewFragment;
 import com.example.b07group7project.shopping_cart.ShoppingCart;
@@ -28,7 +29,7 @@ public class ShopperNavigationActivity extends Navigation {
 
         homeFragment = ShopperViewStoreFragment.newInstance();
         cartFragment = ShoppingCart.newInstance();
-        accountFragment = ItemPreviewFragment.newInstance(); // CHANGE THIS
+        accountFragment = CreateProductFragment.newInstance(); // CHANGE THIS
 
         replaceFragment(homeFragment, false);
